Created list of available MiniLinks created by the user.

+++ src/main/java/locke/dustin/minilink/service/MiniLinkService.java
@@ -10,6 +10,7 @@ import locke.dustin.minilink.util.exception.ExistingUrlException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.UUID;
@@ -86,6 +87,15 @@ public class MiniLinkService {
     }
 
     public void deleteById ( Long id ) {
-        repo.deleteById( id );
+        System.out.println( "MiniLink id: " + id );
+        if (!repo.existsById(id)) {
+            System.out.println( "Cannot find MiniLink with id: " + id );
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    "Link not found"
+            );
+        }
+
+        repo.deleteById(id);
     }
 }
